Update residential.py
The change in the lighting load function is necessary since in the original code the lighting was adapted when the occupants were asleep or away.
The standby-use was only included one time-step before the appliance was turned on, otherwise it was zero. The removal of the indent fixes this problem.